Connect reader to parser and Neo4j
